payment-service/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
django-cors-headers==4.0.0
django-filter==23.2
requests==2.28.2
gunicorn==20.1.0 